requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
PyJWT>=2.8.0
bcrypt>=4.0.0
opencv-python>=4.8.0
numpy>=1.24.0
onnxruntime>=1.16.0
